requirements.txt
absl-py==2.0.0
attrs==23.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
numpy==1.26.3
ortools==9.8.3296
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
protobuf==4.25.1
PyMySQL==1.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
